requirements.txt
Django>=5.0,<6.0
Pillow>=10.0
paynow
requests>=2.31
django-unfold

